int in conflict with persistence LONg

+++ src/main/scala/app/domain/Id.scala
@@ -1,5 +1,5 @@
 package app.domain
 
 abstract class Id {
-  def id: Option[Int]
+  def id: Option[Long]
 }
