+++ src/main/scala/app/domain/third/Third.scala
@@ -2,7 +2,7 @@ package app.domain.third
 
 import app.domain.{Id, RandomTextGenerator}
 
-case class Third(id: Option[Int], val profile: ThirdProfile, val credentials: ThirdCredentials) extends Id with RandomTextGenerator {
+case class Third(id: Option[Long], val profile: ThirdProfile, val credentials: ThirdCredentials) extends Id with RandomTextGenerator {
 
   def refreshCredentials(): Third = {
     val updatedCredentials = credentials.copy(
