+++ src/test/scala/builders/BuildThird.scala
@@ -4,7 +4,7 @@ import app.domain.third.{Third, ThirdCredentials, ThirdProfile}
 
 object BuildThird {
     def any(
-             withId: Option[Int] = Faker(Some(Faker.number()), None),
+             withId: Option[Long] = Faker(Some(Faker.number()), None),
              withThirdProfile: ThirdProfile = BuildThirdProfile.any(),
              withCredentials: ThirdCredentials = BuildThirdCredentials.any()
         ): Third = {
